Module cairo_native::docs::section08

MLIR Resources

How MLIR Works

MLIR is composed of dialects, which is like a IR of it’s own, and this IR can be converted to another dialect IR (if the functionality exists). This is what makes MLIR shine.

Some commonly used dialects in this project:

The arith dialect: It contains arithmetic operations, such as addi, subi for addition and subtraction.

The cf dialect: It contains basic control flow operations, such as the br and cond_br, which are unconditional and conditional jumps.

The IR

The MLIR IR is composed recursively like this: Operation -> Region -> Block -> Operations

Each operation has 1 or more region, each region has 1 or more blocks, each block has 1 or more operations.

This way a MLIR program can be composed.

Transformations and passes

MLIR provides a set of transformations that can optimize the IR. Such as canonicalize.

Check out https://mlir.llvm.org/docs/Canonicalization/ and https://mlir.llvm.org/docs/Passes/.

Translating

In our case, llvm is our target, so we end up translating all dialects down to the LLVM dialect, which then gets converted to LLVM IR.

Learning Resources

Resources marked with → are best.

Introduction

→ 2019 EuroLLVM Developers’ Meeting: MLIR: Multi-Level Intermediate Representation Compiler Infrastructure

→ MLIR: A Compiler Infrastructure for the End of Moore’s Law The paper introducing the MLIR framework

7-minute video summary of paper: Read a paper: Multi-level Intermediate Representation (MLIR)

Another version of the paper: MLIR: Scaling Compiler Infrastructure for Domain Specific Computation

MLIR Tutorial

→ (slides) MLIR Tutorial (LLVM Dev Mtg, 2020)

→ (video) 2020 LLVM Developers’ Meeting: M. Amini & R. Riddle “MLIR Tutorial”

(older slides) MLIR Tutorial (LLVM Developers Meeting, Euro-LLVM 2019)

(older slides) MLIR Tutorial (MLIR 4 HPC, 2019)

(older video) 2019 EuroLLVM Developers’ Meeting: Mehdi & Vasilache & Zinenko “Building a Compiler with MLIR”

→ Another MLIR Tutorial https://github.com/j2kun/mlir-tutorial

→ How to build a compiler with LLVM and MLIR

Other articles, posts

→ Intro to LLVM and MLIR with Rust and Melior

→ MLIR Notes

→ Compilers and IRs: LLVM IR, SPIR-V, and MLIR HN

MLIR: Redefining the compiler infrastructure

Pinch: Implementing a borrow-checked language with MLIR

Official Documentation

MLIR Homepage

MLIR Language Reference

MLIR Compiler Youtube Channel

Talks, Presentations, & Videos

2020 LLVM in HPC Workshop: Keynote: MLIR: an Agile Infrastructure for Building a Compiler Ecosystem

2021 LLVM Dev Mtg “Representing Concurrency with Graph Regions in MLIR”

2022 LLVM Dev Mtg: Paths towards unifying LLVM and MLIR

2022 LLVM Dev Mtg: VAST: MLIR for program analysis of C/C++

2022 LLVM Dev Mtg: MLIR for Functional Programming

2022 EuroLLVM Dev Mtg “Prototyping a Compiler for Homomorphic Encryption Using MLIR”

cirgen: MLIR based compiler for zk-STARK circuit generation - Frank Laub (RISC Zero)

Prototyping a compiler for homomorphic encryption using MLIR

Slides

The HEIR Compiler w/ Jeremy Kun

Useful code

raviqqe/melior

femtomc/mlir-sys Rust bindings to the MLIR C API.

GetFirefly/firefly An alternative BEAM implementation, designed for WebAssembly.