Module cairo_native::easy

The easy to use API to compile and execute cairo programs using the MLIR JIT.

Check out the main crate docs for more information.

Enums

Error

The possible errors encountered when calling compile_and_execute

Functions

compile_and_execute

Shortcut to compile and execute a program.

create_compiler

Creates all the structures needed to compile and create the JIT engine.

create_engine

Creates the execution engine, with all symbols registered.

felt252_bigint

Parse any type that can be a bigint to a felt that can be used in the cairo-native input.

felt252_short_str

Parse a short string into a felt that can be used in the cairo-native input.

felt252_str

Parse a numeric string into felt, wrapping negatives around the prime modulo.

find_entry_point

Returns the given entry point.

get_required_initial_gas

Returns the required initial gas, also inserts de Gas metadata if needed.

run_passes

Runs the needed MLIR passes on the module to execute with the JIT.