cairo_native::error

Module panic

In Cairo Native we want to avoid the use of panic, even in situation where it should never happen. The downside of this is that we lose:

Possible compiler opitimizations

Stack backtrace on error

This modules aims to avoid panics while still obtaining a stack backtrace on eventual errors.

Structs

NativeAssertError

NativeAssertError acts as a non-panicking alternative to Rust’s panic. When the error is created the backtrace or location is captured, which is useful for debugging.

Traits

ToNativeAssertError

Extension trait used to easly convert Results and Options to NativeAssertError