Module cairo_native::executor::contract

A specialized executor for Starknet contracts, avoiding the overhead of storing the sierra program registry and enabling efficient serialization of the program/data once compiled.

This executor heavily relies on the stability of the contract entry point argument order.

Right now this order is like the following:

One or more builtins (rangecheck, etc)

Gas builtin

System builtin (the syscall handler)

A Array of felts (calldata)

How it works:

The only variable data we need to know at call time is the builtins order, to save this, when first compiling the sierra program (with ContractExecutor::new) it iterates through all the user defined functions (this includes the contract wrappers, which are the ones that matter) and saves the builtin arguments in a Vec.

The API provides two more methods: ContractExecutor::save and ContractExecutor::load.

Save can be used to save the compiled program into the given path, alongside it will be saved a json file with the entry points and their builtins (as seen in the example)

{"0":{"builtins":[]},"1":{"builtins":["RangeCheck","Gas","System"]}}

If the given path is “program.so”, then at the same location, “program.json” will be saved.

When loading, passing the “program.so” path will make it load the program and the “program.json” alongside it.

Structs

ContractExecutor

Please look at the module [crate::executor::ce] level docs.

EntryPointInfo

Enums

BuiltinType