Module cairo_native::libfuncs

Compiler libfunc infrastructure

Contains libfunc generation stuff (aka. the actual instructions).

Modules

ap_tracking

AP tracking libfuncs

array

Array libfuncs

bitwise

Bitwise libfuncs

bool

Boolean libfuncs

box

Box libfuncs

branch_align

Branch alignment libfunc

cast

Casting libfuncs

debug

Debug libfuncs

drop

AP tracking libfuncs

dup

State value duplication libfunc

ec

Elliptic curve libfuncs

enum

Enum-related libfuncs

felt252

felt252-related libfuncs

felt252_dict

felt252 dictionary libfuncs

felt252_dict_entry

felt252 dictionary entry libfuncs

function_call

Function call libfuncs

gas

Gas management libfuncs

mem

Memory-related libfuncs

nullable

Nullable libfuncs

pedersen

Pedersen hashing libfuncs

poseidon

Poseidon hashing libfuncs

snapshot_take

Snapshot taking libfuncs

stark_net

StarkNet libfuncs

struct

Struct-related libfuncs

uint8

u8-related libfuncs

uint16

u16-related libfuncs

uint32

u32-related libfuncs

uint64

u64-related libfuncs

uint128

u128-related libfuncs

uint256

u256-related libfuncs

uint512

u512-related libfuncs

unconditional_jump

Unconditional jump libfunc

unwrap_non_zero

Non-zero unwrapping libfuncs

Structs

LibfuncHelper

Helper struct which contains logic generation for extra MLIR blocks and branch operations to the next statements.

Enums

BranchTarget

A libfunc branching target.

Traits

LibfuncBuilder

Generation of MLIR operations from their Sierra counterparts.