cairo_native

Module metadata

Code generation metadata

The metadata is used for various stuff that would be otherwise difficult or plain impossible with the current design, such as:

Pass compile-time constants that would otherwise have to be hardcoded in various places (ex. PrimeModuloMeta).

Declare FFI bindings to external libraries (ex. ReallocBindingsMeta).

Pass extra compilation info to the libfunc generators (ex. TailRecursionMeta).

Modules

auto_breakpoint

debug_utils

Debug utilities

drop_overrides

Dropping logic overrides

dup_overrides

Duplication logic overrides

enum_snapshot_variants

felt252_dict

gas

This file contains the gas calculation metadata.

profiler

The libfunc profiling feature is used to generate information about every libfunc executed in a sierra program.

realloc_bindings

Memory allocation external bindings

runtime_bindings

Runtime library bindings

tail_recursion

Tail recursion information

trace_dump

Structs

MetadataStorage

Metadata container.