Module cairo_native::metadata::drop_overrides

Dropping logic overrides

By default, values are discarded, but some cases (like arrays, boxes, nullables, dictionaries and some structs and enums) need a drop implementation instad. This metadata is a register of types that require a drop implementation as well as the logic to register and invoke those implementations.

Drop implementations

The drop logic is implemented as a function for each type that requires it. It has to be a function to allow self-referencing types. If we inlined the drop implementations, self-referencing types would generate infinite code thus overflowing the stack when generating code.

The generated functions are not public (they are internal) and follow this naming convention:

drop${type id}

where {type id} is the numeric value of the ConcreteTypeId.

Structs

DropOverridesMeta