cairo_native::metadata
Module dup_overrides

Duplication logic overrides

By default, values are copied (aka. memcpy’d), but some cases (like arrays, boxes, nullables, dictionaries and some structs and enums) need a clone implementation instad. This metadata is a register of types that require a clone implementation as well as the logic to register and invoke those implementations.

Clone implementations

The clone logic is implemented as a function for each type that requires it. It has to be a function to allow self-referencing types. If we inlined the clone implementations, self-referencing types would generate infinite code thus overflowing the stack when generating code.

The generated functions are not public (they are internal) and follow this naming convention:

dup${type id}

where {type id} is the numeric value of the ConcreteTypeId.

Structs

DupOverridesMeta