cairo_native::metadata

Module gas

This file contains the gas calculation metadata.

Each statement has an associated GasCost, which represents the cost of executing that statement, in terms of tokens.

To calculate the actual cost, the amount of tokens is multiplied by the cost of the given token type. The cost of each token type is specified on runtime, with the BuiltinCosts structure.

When implementing libfuncs, the GasCost metadata entry already contains the GasCost for the current sierra statement

Structs

GasCost

The gas cost associated to a determined sierra statement.

GasMetadata

Holds global gas info.

StatementCostInfo

Enums

GasMetadataError

Error for metadata calculations.