cairo_native::metadata

Module profiler

The libfunc profiling feature is used to generate information about every libfunc executed in a sierra program.

When this feature is used, the compiler will call the important methods:

measure_timestamp: called before every libfunc execution.

push_frame: called before every branching operation. This method will also call measure_timestamp. This, with the timestamp calculated before the execution, will allow to measure each statement’s execution time. If for some reason, the statement delta time could not be gathered, we just record an unit value, recording that we executed the given statement.

Once the program execution finished and the information was gathered, the get_profile method can be called. It groups the samples by libfunc, and returns all data related to each libfunc.

As well as with the trace-dump feature, in the context of Starknet contracts, we need to add support for building profiles for multiple executions. To do so, we need two important elements, which must be set before every contract execution:

A global static hashmap to map every profile ID to its respective profiler. See LIBFUNC_PROFILE.

A counter to track the ID of the current profiler, which gets updated every time we switch to another contract. Since a contract can call other contracts, we need a way of restoring the counter after every execution.

See cairo-native-run for an example on how to do it.

Structs

LibfuncProfileData

Represents the profile data for a particular libfunc.

ProfilerImpl

ProfilerMeta

Enums

ProfilerBinding

Statics

LIBFUNC_PROFILE

Functions

setup_runtime