cairo_native::metadata

Module realloc_bindings

Memory allocation external bindings

This metadata ensures that the bindings to the C function realloc exist in the current compilation context.

Structs

ReallocBindingsMeta

Memory allocation realloc metadata.