Module cairo_native::types::array

Array type

An array type is a dynamically allocated list of items.

Layout

Being dynamically allocated, we just need to keep the pointer to the data, its length and its capacity:

Index	Type	Description

0	!llvm.ptr<T>	Pointer to the data1.

1	i32	Array length2.

2	i32	Allocated capacity2.

When capacity is zero, this field is not guaranteed to be valid. 1: Both numbers are number of items, not bytes. ↩

↩

Functions

build

Build the MLIR type.