cairo_native

Module utils

Various utilities

Modules

mem_tracing

safe_runner

Safe runner

testing

trace_dump

The trace dump feature is used to generate the execution trace of a sierra program.

walk_ir

Structs

BuiltinCosts

Represents the gas cost of each cost token type

LayoutError

Constants

SHARED_LIBRARY_EXT

Statics

HALF_PRIME

PRIME

The felt252 prime modulo.

Functions

create_engine

Creates the execution engine, with all symbols registered.

debug_with

Return a type that calls a closure when formatted using Debug.

decode_error_message

Decode an UTF-8 error message replacing invalid bytes with their hexadecimal representation, as done by Python’s x.decode('utf-8', errors='backslashreplace').

felt252_bigint

Parse any type that can be a bigint to a felt that can be used in the cairo-native input.

felt252_short_str

Parse a short string into a felt that can be used in the cairo-native input.

felt252_str

Parse a numeric string into felt, wrapping negatives around the prime modulo.

find_entry_point

Returns the given entry point if present.

find_entry_point_by_idx

Returns the given entry point if present.

find_function_id

Given a string representing a function name, searches in the program for the id corresponding to said function, and returns a reference to it.

generate_function_name

Generate a function name.

get_integer_layout

Return the layout for an integer of arbitrary width.

get_types_total_size

Returns the total layout size for the given types.

layout_repeat

Copied from std.

mangle_name

Mangles the given function name to ensure safe compilation.

padding_needed_for

Edit: Copied from the std lib.

run_pass_manager