cairo_native::utils
Module safe_runner

Safe runner

The safe runner provides a way to run contracts without any risk of crashing due to invalid memory accesses, including stack overflows. The same mechanism can also be used to abort the current Cairo program execution, although it will probably leak some memory.

It works by setting a signal handler for the SIGSEGV signal, which is generated by the operating system on invalid memory accesses. This signal handler is global, therefore it must be set by the application, rather than the library, by invoking setup_safe_runner.

Since returning from a SIGSEGV signal handler would crash the program, the safe runner will perform a non-local goto using the setjmp and longjmp functions. The first one configures the jump target while the second performs the goto.

Note: The protection is not 100% guaranteed. Since all memory addresses are within the same memory space, some combinations of base addresses plus offsets may end up in a different page allocation.

Enums

SafeRunnerError

Functions

abort_safe_runner

Manually trigger the segfault handler, thus aborting the current program.

run_safely

Run a closure within a safe runner.

setup_safe_runner

Configure the current process for the [SafeRunner].