Trait cairo_native::utils::RangeExt

pub trait RangeExt {
    // Required methods
    fn zero_based_bit_width(&self) -> u32;
    fn offset_bit_width(&self) -> u32;
}

Required Methods

fn zero_based_bit_width(&self) -> u32

Width in bits when the offset is zero (aka. the natural representation).

fn offset_bit_width(&self) -> u32

Width in bits when the offset is not necessarily zero (aka. the compact representation).