Trait cairo_native::values::ValueBuilder

pub trait ValueBuilder<TType, TLibfunc>where
    TType: GenericType<Concrete = Self>,
    TLibfunc: GenericLibfunc,{
    type Deserializer<'a>: ValueDeserializer<'a, TType, TLibfunc>;
    type Serializer<'a>: ValueSerializer<'a, TType, TLibfunc>;
    type Error: Error;

    // Required methods
    fn is_complex(&self) -> bool;
    unsafe fn debug_fmt(
        &self,
        f: &mut Formatter<'_>,
        id: &ConcreteTypeId,
        registry: &ProgramRegistry<TType, TLibfunc>,
        ptr: NonNull<()>
    ) -> Result;
}

The ValueBuilder trait is implemented any de/serializable value, which is the TType generic.

Required Associated Types

type Deserializer<'a>: ValueDeserializer<'a, TType, TLibfunc>

Value deserializer from serde into the JIT ABI.

type Serializer<'a>: ValueSerializer<'a, TType, TLibfunc>

Value serializer from the JIT ABI into serde.

type Error: Error

Error type returned from the de/serializers.

Required Methods

fn is_complex(&self) -> bool

Return whether the type is considered complex or simple.

Complex types are always passed by pointer (both as params and return values) and require a stack allocation. Examples of complex values include structs and enums, but not felts since LLVM considers them integers.

unsafe fn debug_fmt( &self, f: &mut Formatter<'_>, id: &ConcreteTypeId, registry: &ProgramRegistry<TType, TLibfunc>, ptr: NonNull<()> ) -> Result

Write a representation of the argument in a human-friendly format for debugging.