Module cairo_native::executor

Executors

This module provides methods to execute the programs, either via JIT or compiled ahead of time. It also provides a cache to avoid recompiling previously compiled programs.

Structs

AotContractExecutor

Please look at the module level docs.

AotNativeExecutor

JitNativeExecutor

A MLIR JIT execution engine in the context of Cairo Native.