Function cairo_native::compile

pub fn compile(
    context: &Context,
    module: &Module<'_>,
    program: &Program,
    registry: &ProgramRegistry<CoreType, CoreLibfunc>,
    metadata: &mut MetadataStorage,
    di_compile_unit_id: Attribute<'_>,
    ignore_debug_names: bool,
) -> Result<(), Error>

Run the compiler on a program. The compiled program is stored in the MLIR module.

The generics TType and TLibfunc contain the information required to generate the MLIR types and statement operations. Most of the time you’ll want to use the default ones, which are CoreType and CoreLibfunc respectively.

This function needs the program and the program’s registry, which doesn’t need to have AP tracking information.

Additionally, it needs a reference to the MLIR context, the output module and the metadata storage. The last one is passed externally so that stuff can be initialized if necessary.