Module cairo_native::metadata

Code generation metadata

The metadata is used for various stuff that would be otherwise difficult or plain impossible with the current design, such as:

Pass compile-time constants that would otherwise have to be hardcoded in various places (ex. PrimeModuloMeta).

Declare FFI bindings to external libraries (ex. ReallocBindingsMeta).

Pass extra compilation info to the libfunc generators (ex. TailRecursionMeta).

Modules

auto_breakpoint

debug_utils

Debug utilities

drop_overrides

Dropping logic overrides

dup_overrides

Duplication logic overrides

enum_snapshot_variants

gas

realloc_bindings

Memory allocation external bindings

runtime_bindings

Runtime library bindings

tail_recursion

Tail recursion information

Structs

MetadataStorage

Metadata container.