Module cairo_native::types::enum

Enum type

Enumerations are special because they can potentially represent an unlimited amount of things at the same time. They are similar to Rust enums since they can contain data along with the discriminator.

Layout

Index	Type	Description

0	iN	Discriminant.

1	Depends on variants.	Payload.

As seen in the table above, an enum’s layout is not as simple as concatenating the discriminant with the payload.

The discriminant will have the bit width required to store all possible values. The following table contains an example of some number of variants with their discriminant type:

Number of variants	Discriminant type	ABI (in Rust types)

0 or 1	i0	()

2	i1	u8

3 or 4	i2	u8

5, 6, 7 or 8	i3	u8

9 to 16	i4	u8

129 to 256	i8	u8

257 to 512	i9	u16

32769 to 65536	i16	u16

65537 to 131072	i17	u32

In Rust, the number of bits and bytes required can be obtained using the following formula:
n bits = { 0 if n variants = 0 ⌈ log 2 n variants ⌉ if n variants ≠ 0 n bytes = ⌈ n bits 8 ⌉
The payload will then be appended to the discriminant after applying its alignment rules. This will cause unused space between the tag and the payload in most cases. As an example, the following enum will have the layouts described in the table below.

enum MyEnum {
    U8: u8,
    U16: u16,
    U32: u32,
    U64: u64,
    Felt: Felt,
}

MyEnum::U8

Index 	Type 	ABI (in Rust types) 	Alignment 	Size 	Description

0 	i3 	u8 	1 	1 	Discriminant.

1 	i8 	u8 	1 	1 	Payload.

N/A 	N/A 	[u8; 38] 	1 	38 	Padding.

MyEnum::U16

Index 	Type 	ABI (in Rust types) 	Alignment 	Size 	Description

0 	i3 	u8 	1 	1 	Discriminant.

N/A 	N/A 	[u8; 1] 	1 	1 	Padding.

1 	i16 	u16 	2 	2 	Payload.

N/A 	N/A 	[u8; 36] 	1 	36 	Padding.

MyEnum::U32

Index 	Type 	ABI (in Rust types) 	Alignment 	Size 	Description

0 	i3 	u8 	1 	1 	Discriminant.

N/A 	N/A 	[u8; 3] 	1 	3 	Padding.

1 	i32 	u32 	4 	4 	Payload.

N/A 	N/A 	[u8; 32] 	1 	32 	Padding.

MyEnum::U64

Index 	Type 	ABI (in Rust types) 	Alignment 	Size 	Description

0 	i3 	u8 	1 	1 	Discriminant.

N/A 	N/A 	[u8; 7] 	1 	7 	Padding.

1 	i64 	u64 	8 	8 	Payload.

N/A 	N/A 	[u8; 24] 	1 	24 	Padding.

MyEnum

Index 	Type 	ABI (in Rust types) 	Alignment 	Size 	Description

0 	i3 	u8 	1 	1 	Discriminant.

N/A 	N/A 	[u8; 7] 	1 	7 	Padding.

1 	i252 	[u64; 4] 	8 	32 	Payload.

As seen above, while the discriminant is always at the same offset, the payloads don’t necessary have the same offset between all variants. It depends on the payload’s alignment.

In reality, the first variant will have a zero-sized padding between the discriminant and the payload to keep everything consistent and the padding will have its own index, shifting every index below it by one. However all that’s been ignored for documenting purposes.

An MLIR type cannot be an enumeration (it doesn’t exist), therefore a variant or a buffer has to be used. Using a buffer as a dummy payload has been discarded because it doesn’t keep the enum’s alignment information. To keep that info, the first variant with the biggest alignment is used as the default payload.

Using the info stated above, we can infer that the example enum will have the following type by default:

Index	Type	ABI (in Rust types)	Alignment	Size	Description

0	i3	u8	1	1	Discriminant.

N/A	N/A	[u8; 7]	1	7	Padding.

1	i64	u64	8	8	Payload.

N/A	N/A	[u8; 24]	1	24	Padding.

Functions

build

Build the MLIR type.

get_layout_for_variants

Extract layout for the default enum representation, its discriminant and all its payloads.

get_type_for_variants

Extract the type and layout for the default enum representation, its discriminant and all its payloads.

Type Aliases

TypeLayout

An MLIR type with its memory layout.