Trait cairo_native::types::TypeBuilder

pub trait TypeBuilder<TType, TLibfunc>where
    TType: GenericType<Concrete = Self>,
    TLibfunc: GenericLibfunc,{
    type Error: Error;

    // Required methods
    fn build<'ctx>(
        &self,
        context: &'ctx Context,
        module: &Module<'ctx>,
        registry: &ProgramRegistry<TType, TLibfunc>,
        metadata: &mut MetadataStorage,
        self_ty: &ConcreteTypeId
    ) -> Result<Type<'ctx>, Self::Error>;
    fn is_builtin(&self) -> bool;
    fn is_complex(&self, registry: &ProgramRegistry<TType, TLibfunc>) -> bool;
    fn is_zst(&self, registry: &ProgramRegistry<TType, TLibfunc>) -> bool;
    fn layout(
        &self,
        registry: &ProgramRegistry<TType, TLibfunc>
    ) -> Result<Layout, Self::Error>;
    fn is_memory_allocated(
        &self,
        registry: &ProgramRegistry<TType, TLibfunc>
    ) -> bool;
    fn integer_width(&self) -> Option<usize>;
    fn variants(&self) -> Option<&[ConcreteTypeId]>;
    fn build_drop<'ctx, 'this>(
        &self,
        context: &'ctx Context,
        registry: &ProgramRegistry<TType, TLibfunc>,
        entry: &'this Block<'ctx>,
        location: Location<'ctx>,
        helper: &LibfuncHelper<'ctx, 'this>,
        metadata: &mut MetadataStorage,
        self_ty: &ConcreteTypeId
    ) -> Result<(), Self::Error>;
}

Generation of MLIR types from their Sierra counterparts.

All possible Sierra types must implement it. It is already implemented for all the core Sierra types, contained in [CoreTypeConcrete].

Required Associated Types

type Error: Error

Error type returned by this trait’s methods.

Required Methods

fn build<'ctx>( &self, context: &'ctx Context, module: &Module<'ctx>, registry: &ProgramRegistry<TType, TLibfunc>, metadata: &mut MetadataStorage, self_ty: &ConcreteTypeId ) -> Result<Type<'ctx>, Self::Error>

Build the MLIR type.

fn is_builtin(&self) -> bool

Return whether the type is a builtin.

fn is_complex(&self, registry: &ProgramRegistry<TType, TLibfunc>) -> bool

Return whether the type requires a return pointer when returning.

fn is_zst(&self, registry: &ProgramRegistry<TType, TLibfunc>) -> bool

Return whether the Sierra type resolves to a zero-sized type.

fn layout( &self, registry: &ProgramRegistry<TType, TLibfunc> ) -> Result<Layout, Self::Error>

Generate the layout of the MLIR type.

Used in both the compiler and the interface when calling the compiled code.

fn is_memory_allocated( &self, registry: &ProgramRegistry<TType, TLibfunc> ) -> bool

Whether the layout should be allocated in memory (either the stack or the heap) when used as a function invocation argument or return value.

fn integer_width(&self) -> Option<usize>

If the type is an integer (felt not included) type, return its width in bits.

TODO: How is it used?

fn variants(&self) -> Option<&[ConcreteTypeId]>

If the type is a variant type, return all possible variants.

TODO: How is it used?

fn build_drop<'ctx, 'this>( &self, context: &'ctx Context, registry: &ProgramRegistry<TType, TLibfunc>, entry: &'this Block<'ctx>, location: Location<'ctx>, helper: &LibfuncHelper<'ctx, 'this>, metadata: &mut MetadataStorage, self_ty: &ConcreteTypeId ) -> Result<(), Self::Error>