Module cairo_native::utils

Various utilities

Structs

LayoutError

Constants

SHARED_LIBRARY_EXT

Traits

ProgramRegistryExt

Functions

cairo_to_sierra

Compile a cairo program found at the given path to sierra.

create_engine

Creates the execution engine, with all symbols registered.

debug_with

Return a type that calls a closure when formatted using Debug.

felt252_bigint

Parse any type that can be a bigint to a felt that can be used in the cairo-native input.

felt252_short_str

Parse a short string into a felt that can be used in the cairo-native input.

felt252_str

Parse a numeric string into felt, wrapping negatives around the prime modulo.

find_entry_point

Returns the given entry point if present.

find_entry_point_by_idx

Returns the given entry point if present.

find_function_id

Given a string representing a function name, searches in the program for the id corresponding to said function, and returns a reference to it.

generate_function_name

Generate a function name.

get_integer_layout

Return the layout for an integer of arbitrary width.

layout_repeat

Copied from std.

next_multiple_of_u32

next_multiple_of_usize

padding_needed_for

Edit: Copied from the std lib.

register_runtime_symbols

run_pass_manager