cairo_native::utils
Function padding_needed_for

pub const fn padding_needed_for(layout: &Layout, align: usize) -> usize

Edit: Copied from the std lib.

Returns the amount of padding we must insert after layout to ensure that the following address will satisfy align (measured in bytes).

e.g., if layout.size() is 9, then layout.padding_needed_for(4) returns 3, because that is the minimum number of bytes of padding required to get a 4-aligned address (assuming that the corresponding memory block starts at a 4-aligned address).

The return value of this function has no meaning if align is not a power-of-two.

Note that the utility of the returned value requires align to be less than or equal to the alignment of the starting address for the whole allocated block of memory. One way to satisfy this constraint is to ensure align <= layout.align().