Module cairo_native::values

JIT params and return values de/serialization

A Rusty interface to provide parameters to JIT calls.

Enums

JitValue

A JitValue is a value that can be passed to the JIT engine as an argument or received as a result.